requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
setuptools==70.0.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
